backend/requirements-dev.txt
-r requirements.txt
pytest==8.3.2
pytest-flask==1.3.0
python-dotenv==1.0.1
